requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.6
django-ckeditor==6.7.1
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.0
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
